Use hpctoolkit persistent id for operations

+++ src/tool/hpcrun/gpu/nvidia/sanitizer-buffer.c
@@ -78,7 +78,7 @@ typedef struct sanitizer_buffer_t {
   uint32_t thread_id;
   uint32_t cubin_id;
   uint32_t mod_id;
-  uint64_t kernel_id;
+  int32_t kernel_id;
   uint64_t host_op_id;
   gpu_patch_buffer_t *gpu_patch_buffer;
 } sanitizer_buffer_t;
@@ -96,7 +96,7 @@ sanitizer_buffer_process
   uint32_t thread_id = b->thread_id;
   uint32_t cubin_id = b->cubin_id;
   uint32_t mod_id = b->mod_id;
-  uint64_t kernel_id = b->kernel_id;
+  int32_t kernel_id = b->kernel_id;
   uint64_t host_op_id = b->host_op_id;
 
   gpu_patch_buffer_t *gpu_patch_buffer = b->gpu_patch_buffer;
@@ -122,7 +122,7 @@ sanitizer_buffer_produce
  uint32_t thread_id,
  uint32_t cubin_id,
  uint32_t mod_id,
- uint64_t kernel_id,
+ int32_t kernel_id,
  uint64_t host_op_id,
  size_t num_records,
  atomic_uint *balance,
