+++ src/tool/hpcrun/gpu/nvidia/sanitizer-buffer.h
@@ -76,7 +76,7 @@ sanitizer_buffer_produce
  uint32_t thread_id,
  uint32_t cubin_id,
  uint32_t mod_id,
- uint64_t kernel_id,
+ int32_t kernel_id,
  uint64_t host_op_id,
  size_t num_records,
  atomic_uint *balance,
